requirements.txt
certifi
absl-py==1.3.0
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-binary==110.0.5481.30.0
chromedriver-binary-auto==0.2.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
DateTime==4.7
decorator==5.1.1
Deprecated==1.2.13
exceptiongroup==1.1.0
executing==1.2.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
ipython==8.6.0
jedi==0.18.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
mahotas==1.4.13
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
protobuf==3.19.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
selenium==4.8.0
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
tf-keras-vis==0.8.2
threadpoolctl==3.1.0
tifffile==2022.10.10
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
zope.interface==5.5.1
